requirements.txt
openai>=1.0.0
google-generativeai>=0.5.0
google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
pytest>=7.4.0
pdfplumber>=0.11.0
instructor>=1.3.0
